fazendas_api/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
psycopg2-binary
pydantic
python-dotenv
pytest
httpx

# Geo
geopandas
shapely
fiona
pyproj
